Fixed NPE from method being fired from uninitialised class
Moved the method to Util so that it can be used by DwarfEffect whilst
the effects.csv is being loaded by the config manager
